@@ -1,7 +1,6 @@
 import React from "react";
 import { MdOutlineCancel } from "react-icons/md";
 import { BsCheck } from "react-icons/bs";
-import { TooltipComponent } from "@syncfusion/ej2-react-popups";
 import { themeColors } from "../data/content";
 import { useStateContext } from "../contexts/ContextProvider";
 
@@ -10,8 +9,14 @@ const Settings = () => {
     useStateContext();
 
   return (
-    <div onClick={() => setSettings(false)} className="bg-half-transparent w-screen fixed nav-item top-0 right-0">
-      <div onClick={(e) => e.stopPropagation()} className="float-right h-screen dark:text-gray-200 bg-white dark:bg-[#484B52] w-400">
+    <div
+      onClick={() => setSettings(false)}
+      className="bg-half-transparent w-screen fixed nav-item top-0 right-0"
+    >
+      <div
+        onClick={(e) => e.stopPropagation()}
+        className="float-right h-screen dark:text-gray-200 bg-white dark:bg-[#484B52] w-400"
+      >
         <div className="flex justify-between items-center p-4 ml-4">
           <p className="font-semibold text-xl">Settings</p>
           <button
@@ -58,26 +63,23 @@ const Settings = () => {
           <p className="font-semibold text-lg">Theme Colors</p>
           <div className="flex gap-3">
             {themeColors.map((item, index) => (
-              <TooltipComponent
+              <div
                 key={index}
-                content={item.name}
-                position="TopCenter"
+                className="relative mt-2 cursor-pointer flex gap-5 items-center"
               >
-                <div className="relative mt-2 cursor-pointer flex gap-5 items-center">
-                  <button
-                    type="button"
-                    className="h-10 w-10 rounded-full cursor-pointer"
-                    style={{ backgroundColor: item.color }}
-                    onClick={() => setColor(item.color)}
-                  >
-                    <BsCheck
-                      className={`ml-2 text-2xl text-white ${
-                        item.color === currentColor ? "block" : "hidden"
-                      }`}
-                    />
-                  </button>
-                </div>
-              </TooltipComponent>
+                <button
+                  type="button"
+                  className="h-10 w-10 rounded-full cursor-pointer"
+                  style={{ backgroundColor: item.color }}
+                  onClick={() => setColor(item.color)}
+                >
+                  <BsCheck
+                    className={`ml-2 text-2xl text-white ${
+                      item.color === currentColor ? "block" : "hidden"
+                    }`}
+                  />
+                </button>
+              </div>
             ))}
           </div>
         </div>
